Website/requirements.txt
flask
pyserial
mysql-connector-python